途中までのコード上げる
