+++ db/migrate.go
@@ -4,9 +4,9 @@ import (
 	"database/sql"
 
 	_ "github.com/go-sql-driver/mysql"
+	"github.com/golang-migrate/migrate"
 	"github.com/golang-migrate/migrate/database/mysql"
 	_ "github.com/golang-migrate/migrate/source/file"
-	"github.com/golang-migrate/migrate/v4"
 )
 
 func main() {
