+++ infra/persistence/user.go
@@ -5,8 +5,7 @@ import (
 	"github.com/ryota1116/stacked_books/domain/repository"
 	"errors"
 	"fmt"
-	"gorm.io/driver/mysql"
-	"gorm.io/gorm"
+
 )
 
 // Userのインフラ層の構造体
@@ -68,19 +67,4 @@ func (up userPersistence) ShowUser(params map[string]string) model.User {
 	return user
 }
 
-// DBサーバーに接続する
-func DbConnect() (db *gorm.DB) {
-	// DB接続
-	dsn := "root@tcp(127.0.0.1:3306)/stacked_books_development?charset=utf8mb4&parseTime=True&loc=Local"
-	db, err := gorm.Open(mysql.Open(dsn), &gorm.Config{})
 
-	if err != nil {
-		panic(err.Error())
-	}
-
-	// DBにテーブルが存在するか確認（存在すればtrueを返す）
-	dbPresence := db.Migrator().HasTable("users")
-	fmt.Println(dbPresence)
-
-	return db
-}
