+++ usecase/user.go
@@ -1,10 +1,10 @@
 package usecase
 
 import (
-	"github.com/ryota1116/stacked_books/domain/model"
-	"github.com/ryota1116/stacked_books/domain/repository"
 	"fmt"
 	"github.com/dgrijalva/jwt-go"
+	"github.com/ryota1116/stacked_books/domain/model"
+	"github.com/ryota1116/stacked_books/domain/repository"
 	"golang.org/x/crypto/bcrypt"
 	"time"
 )
